Remove max_matches from configuration (#4)
The 'max_matches' setting was permanently removed from the configuration
file in Sphinx 2.2.3-beta. An option with the same name is still
available per-query and it is 1000 by default just like it was in the
default config as well. Thus, it is sufficient to remove the setting
from the configurations in this role.

More information:
https://sphinxsearch.com/docs/manual-2.3.2.html#sphinx-deprecations-defaults
https://sphinxsearch.com/docs/manual-2.3.2.html#sphinxql-select

This patch removes the following reoccurring warning in the logs:

    WARNING: key 'max_matches' was permanently removed from Sphinx
    configuration. Refer to documentation for details.